PLY-66: Split music.kt into 4 focused domain-specific files
- Extract MusicModels.kt (356 lines) - Pure data models and API responses
- Extract MusicDiscovery.kt (676 lines) - pCloud integration and discovery
- Extract MusicLibrary.kt (322 lines) - Cache, indexing, sync services
- Extract MusicSearch.kt (93 lines) - Search, filter, browse functionality
- Update Music.kt (33 lines) - Compatibility layer with migration docs

✅ All tests pass - functionality preserved
✅ Clear dependency hierarchy: models → discovery → library → search
✅ Improved code navigation and development experience
✅ Reduced from 1,400+ line monolith to focused ~350 line domains
