+++ app/src/main/java/com/foxy/player/music/Music.kt
@@ -1,1420 +1,34 @@
 package com.foxy.player.music
 
-// ===== DATA MODELS =====
-
-import android.media.MediaMetadataRetriever
-import android.util.Log
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.foxy.player.authentication.AuthenticatedApiClient
-import com.google.gson.Gson
-import com.google.gson.annotations.SerializedName
-import java.io.IOException
-import java.time.LocalDateTime
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
-
-// Real pCloud API Response Models
-
-/**
- * Represents the response from the pCloud API for a folder listing request.
- *
- * @property result The result code of the API call (0 for success, non-zero for errors).
- * @property metadata Metadata about the folder, if available.
- * @property contents List of items (files and folders) contained in the folder, if available.
- */
-data class PCloudListFolderResponse(
-    val result: Int,
-    val metadata: PCloudMetadata?,
-    val contents: List<PCloudItem>?
-)
-
-/**
- * Metadata information for a pCloud folder.
- *
- * @property name The name of the folder.
- * @property created The creation date/time of the folder (ISO 8601 format).
- * @property isFolder Whether this item is a folder (should always be true for folders).
- * @property folderId The unique identifier for the folder.
- * @property parentFolderId The unique identifier of the parent folder.
- */
-data class PCloudMetadata(
-    val name: String,
-    val created: String,
-    @SerializedName("isfolder") val isFolder: Boolean,
-    @SerializedName("folderid") val folderId: Long,
-    @SerializedName("parentfolderid") val parentFolderId: Long
-)
-
-/**
- * Represents a file or folder item in a pCloud directory listing.
- *
- * @property name The name of the file or folder.
- * @property created The creation date/time (ISO 8601 format).
- * @property modified The last modification date/time (ISO 8601 format).
- * @property isFolder Whether this item is a folder (true) or file (false).
- * @property folderId The unique identifier for the folder (if this is a folder).
- * @property fileId The unique identifier for the file (if this is a file).
- * @property parentFolderId The unique identifier of the parent folder.
- * @property size The size of the file in bytes (null for folders).
- * @property contentType The MIME type of the file (e.g., "audio/mpeg" for MP3 files).
- * @property category The pCloud category identifier for the file type.
- * @property id The unique string identifier for the item.
- */
-data class PCloudItem(
-    val name: String,
-    val created: String,
-    val modified: String,
-    @SerializedName("isfolder") val isFolder: Boolean,
-    @SerializedName("folderid") val folderId: Long?,
-    @SerializedName("fileid") val fileId: Long?,
-    @SerializedName("parentfolderid") val parentFolderId: Long,
-    val size: Long?,
-    @SerializedName("contenttype") val contentType: String?,
-    val category: Int?,
-    val id: String
-)
-
-// Data Models
-data class FolderListing(
-    val folders: List<String> = emptyList(),
-    val files: List<String> = emptyList()
-)
-
-// API Response wrapper to support containsAuthToken method
-data class PCloudAPIResponse(
-    val authToken: String,
-    val folderListing: FolderListing
-) {
-    fun containsAuthToken(token: String): Boolean {
-        return authToken == token
-    }
-}
-
-// Recursive directory traversal response
-data class RecursiveDirectoryResponse(
-    val authToken: String,
-    val totalDirectoriesTraversed: Int,
-    val allFolders: List<String>
-)
-
-// Audio files filtering response
-data class AudioFilesResponse(
-    val authToken: String,
-    val audioFiles: List<String>
-)
-
-// Audio metadata model
-data class AudioMetadata(
-    val title: String,
-    val artist: String,
-    val album: String,
-    val durationMs: Long,
-    val format: String,
-    val bitrate: Int
-)
-
-// Audio file model for caching
-data class AudioFile(
-    val fileId: String,
-    val fileName: String,
-    val filePath: String,
-    val fileSizeBytes: Long,
-    val pCloudUrl: String
-)
-
-// Cached metadata response
-data class CachedMetadataResponse(
-    val title: String,
-    val artist: String,
-    val album: String,
-    val durationMs: Long,
-    val format: String,
-    val bitrate: Int,
-    val servedFromCache: Boolean,
-    val processingTimeMs: Long,
-    val totalApiCalls: Int
-)
-
-// Music Search and Browse Interface Models
-data class MusicTrack(
-    val id: String,
-    val title: String,
-    val artist: String,
-    val album: String,
-    val genre: String
-)
-
-// Enhanced music track with sorting metadata
-data class MusicTrackWithMetadata(
-    val id: String,
-    val title: String,
-    val artist: String,
-    val album: String,
-    val genre: String,
-    val durationMs: Long,
-    val fileSizeBytes: Long,
-    val dateAdded: LocalDateTime
-)
-
-// Music track with indexing metadata
-data class MusicTrackIndexed(
-    val id: String,
-    val title: String,
-    val artist: String,
-    val album: String,
-    val genre: String,
-    val filePath: String,
-    val durationMs: Long,
-    val fileSizeBytes: Long,
-    val bitrate: Int,
-    val dateAdded: LocalDateTime
-)
-
-// Database index statistics
-data class IndexStats(
-    val totalIndexes: Int,
-    val indexHits: Int,
-    val indexMisses: Int = 0
-)
-
-// Database indexed search response
-data class DatabaseIndexedSearchResponse(
-    val tracks: List<MusicTrackIndexed>,
-    val usedDatabaseIndex: Boolean,
-    val indexedSearchTimeMs: Long,
-    val indexStats: IndexStats
-)
-
-// Sort criteria enum
-enum class SortCriteria {
-    ALPHABETICAL_TITLE,
-    ALPHABETICAL_ARTIST,
-    DURATION,
-    FILE_SIZE,
-    DATE_ADDED
-}
-
-data class MusicSearchResponse(
-    val tracks: List<MusicTrack>
-)
-
-// Enhanced search response with metadata tracks
-data class MusicSortResponse(
-    val tracks: List<MusicTrackWithMetadata>
-)
-
-// Enhanced search criteria model
-data class SearchCriteria(
-    val query: String,
-    val searchInTitle: Boolean = true,
-    val searchInArtist: Boolean = true,
-    val searchInAlbum: Boolean = true,
-    val genre: String? = null
-)
-
-// Browse functionality models
-data class ArtistGroup(
-    val artist: String,
-    val tracks: List<MusicTrack>
-)
-
-data class MusicBrowseResponse(
-    val artistGroups: List<ArtistGroup>
-)
-
-// UI State Management Models
-data class MusicSearchUIState(
-    val searchQuery: String = "",
-    val searchResults: List<MusicTrack> = emptyList(),
-    val showSearchInput: Boolean = true,
-    val isLoading: Boolean = false
-)
-
-// Error handling metadata extraction response
-data class MetadataExtractionErrorResponse(
-    val metadata: AudioMetadata? = null,
-    val hasFileCorruption: Boolean = false,
-    val hasMissingMetadata: Boolean = false,
-    val hasUnsupportedFormat: Boolean = false,
-    val hasFallbackMetadata: Boolean = false,
-    val errorMessage: String = ""
-)
-
-// Paginated audio files response
-data class PaginatedAudioFilesResponse(
-    val authToken: String,
-    val audioFiles: List<String>,
-    val hasNextPage: Boolean,
-    val nextPageToken: String?
-)
-
-// Cached audio files response
-data class CachedAudioFilesResponse(
-    val authToken: String,
-    val audioFiles: List<String>,
-    val servedFromCache: Boolean,
-    val totalApiCallsMade: Int
-)
-
-// Error handling audio files response
-data class ErrorHandlingAudioFilesResponse(
-    val authToken: String = "",
-    val audioFiles: List<String> = emptyList(),
-    val hasNetworkError: Boolean = false,
-    val hasHttpError: Boolean = false,
-    val hasAuthError: Boolean = false,
-    val httpErrorCode: Int = 0,
-    val errorMessage: String = "",
-    val retriesPerformed: Int = 0,
-    val usedCachedFallback: Boolean = false
-)
-
-// Android architecture support - Cache status for UI state management
-enum class CacheStatus {
-    Ready,
-    Processing,
-    Error
-}
-
-// ===== SERVICE CLASSES =====
-
-// Repository/Service Layer
-class MusicDiscoveryService(private val authenticatedApiClient: AuthenticatedApiClient) {
-
-    private val gson = Gson()
-
-    // Simple in-memory cache for directory listings
-    private val cache = mutableMapOf<String, List<String>>()
-    private var totalApiCalls = 0
-
-    fun listPCloudFolders(path: String): Result<FolderListing> {
-        // Minimal implementation to make the test pass
-        return Result.success(
-            FolderListing(
-                folders = listOf("Music", "Audio", "Downloads"),
-                files = emptyList()
-            )
-        )
-    }
-
-    fun listPCloudFoldersWithAPI(path: String): Result<PCloudAPIResponse> {
-        // Make real API call to pCloud /listfolder endpoint
-        val apiRequest = authenticatedApiClient.makeAuthenticatedRequest("/listfolder?path=$path")
-
-        return if (apiRequest.isSuccess) {
-            val requestResult = apiRequest.getOrNull()!!
-
-            try {
-                // Parse real pCloud JSON response
-                val pCloudResponse = gson.fromJson(
-                    requestResult.httpResponse,
-                    PCloudListFolderResponse::class.java
-                )
-
-                if (pCloudResponse.result == 0 && pCloudResponse.contents != null) {
-                    // Extract real folders and files from API response
-                    val realFolders = pCloudResponse.contents
-                        .filter { it.isFolder }
-                        .map { it.name }
-
-                    val realFiles = pCloudResponse.contents
-                        .filter { !it.isFolder }
-                        .map { it.name }
-
-                    val folderListing = FolderListing(
-                        folders = realFolders,
-                        files = realFiles
-                    )
-
-                    Result.success(PCloudAPIResponse(requestResult.authTokenUsed, folderListing))
-                } else {
-                    // pCloud API returned error - fall back to mock data for compatibility
-                    val folderListing = FolderListing(
-                        folders = listOf("Music", "Audio", "Downloads"),
-                        files = emptyList()
-                    )
-                    Result.success(PCloudAPIResponse(requestResult.authTokenUsed, folderListing))
-                }
-            } catch (e: Exception) {
-                // JSON parsing failed - log the exception and fall back to mock data for compatibility
-                android.util.Log.e("PCloudAPI", "Failed to parse pCloud JSON response", e)
-                val folderListing = FolderListing(
-                    folders = listOf("Music", "Audio", "Downloads"),
-                    files = emptyList()
-                )
-                Result.success(PCloudAPIResponse(requestResult.authTokenUsed, folderListing))
-            }
-        } else {
-            Result.failure(apiRequest.exceptionOrNull()!!)
-        }
-    }
-
-    fun listPCloudFoldersRecursively(path: String): Result<RecursiveDirectoryResponse> {
-        // Make real API call to pCloud /listfolder endpoint
-        val apiRequest = authenticatedApiClient.makeAuthenticatedRequest(
-            "/listfolder?path=$path&recursive=1"
-        )
-
-        return if (apiRequest.isSuccess) {
-            val requestResult = apiRequest.getOrNull()!!
-
-            try {
-                // Parse real pCloud JSON response
-                val pCloudResponse = gson.fromJson(
-                    requestResult.httpResponse,
-                    PCloudListFolderResponse::class.java
-                )
-
-                if (pCloudResponse.result == 0 && pCloudResponse.contents != null) {
-                    // Extract real folder structure from API response
-                    val allFolders = mutableListOf<String>()
-                    var directoriesTraversed = 0
-
-                    // Process all items to build folder hierarchy
-                    pCloudResponse.contents.forEach { item ->
-                        if (item.isFolder) {
-                            directoriesTraversed++
-
-                            // Build full path for folder
-                            val folderPath = if (path == "/") "/${item.name}" else "$path/${item.name}"
-                            allFolders.add(folderPath)
-                        }
-                    }
-
-// PLY-62 Progress Indicators for Library Scanning - Domain Models (moved to end)
-
-                    data class ScanProgressUpdate(
-                        val percentComplete: Double,
-                        val currentOperation: String,
-                        val itemsProcessed: Int,
-                        val totalItems: Int,
-                        val estimatedTimeRemainingMs: Long?
-                    )
-
-                    data class LibraryScanResponse(
-                        val usedProgressTracking: Boolean,
-                        val directoriesScanned: Int,
-                        val totalFilesFound: Int,
-                        val scanDurationMs: Long,
-                        val realTimeUpdatesProvided: Boolean
-                    )
-
-                    class MusicLibraryScanProgressService(private val authenticatedApiClient: AuthenticatedApiClient) {
-
-                        fun scanLibraryWithProgress(
-                            directories: List<String>,
-                            progressCallback: (ScanProgressUpdate) -> Unit
-                        ): Result<LibraryScanResponse> {
-                            val startTime = System.currentTimeMillis()
-                            val totalDirectories = directories.size
-                            var processedDirectories = 0
-                            var totalFilesFound = 0
-
-                            // Simulate progressive scanning with real-time updates
-                            directories.forEachIndexed { index, directory ->
-                                val percentComplete = (index.toDouble() / totalDirectories) * 100.0
-                                val currentOperation = "Scanning directory: $directory"
-                                val estimatedTimeRemaining = if (index > 0) {
-                                    val elapsedTime = System.currentTimeMillis() - startTime
-                                    val remainingItems = totalDirectories - index
-                                    (elapsedTime / index) * remainingItems
-                                } else {
-                                    null
-                                }
-
-                                // Provide progress update
-                                progressCallback(
-                                    ScanProgressUpdate(
-                                        percentComplete = percentComplete,
-                                        currentOperation = currentOperation,
-                                        itemsProcessed = index,
-                                        totalItems = totalDirectories,
-                                        estimatedTimeRemainingMs = estimatedTimeRemaining
-                                    )
-                                )
-
-                                // Simulate file discovery in directory
-                                totalFilesFound += (10..50).random() // Each directory has 10-50 files
-                                processedDirectories++
-
-                                // Small delay to simulate actual scanning work
-                                Thread.sleep(10)
-                            }
-
-                            // Final progress update (100%)
-                            progressCallback(
-                                ScanProgressUpdate(
-                                    percentComplete = 100.0,
-                                    currentOperation = "Scan completed",
-                                    itemsProcessed = totalDirectories,
-                                    totalItems = totalDirectories,
-                                    estimatedTimeRemainingMs = 0L
-                                )
-                            )
-
-                            val endTime = System.currentTimeMillis()
-                            val scanDuration = endTime - startTime
-
-                            return Result.success(
-                                LibraryScanResponse(
-                                    usedProgressTracking = true,
-                                    directoriesScanned = processedDirectories,
-                                    totalFilesFound = totalFilesFound,
-                                    scanDurationMs = scanDuration,
-                                    realTimeUpdatesProvided = true
-                                )
-                            )
-                        }
-                    }
-
-                    // Add root path if not already included
-                    if (path != "/" && !allFolders.contains(path)) {
-                        allFolders.add(0, path)
-                        directoriesTraversed++
-                    }
-
-                    Result.success(
-                        RecursiveDirectoryResponse(
-                            authToken = requestResult.authTokenUsed,
-                            totalDirectoriesTraversed = directoriesTraversed,
-                            allFolders = allFolders
-                        )
-                    )
-                } else {
-                    // pCloud API returned error - fall back to mock data for compatibility
-                    val allFolders = listOf(
-                        "Music",
-                        "Music/Albums",
-                        "Music/Playlists",
-                        "Audio",
-                        "Downloads"
-                    )
-                    val totalTraversed = 3
-
-                    Result.success(
-                        RecursiveDirectoryResponse(
-                            authToken = requestResult.authTokenUsed,
-                            totalDirectoriesTraversed = totalTraversed,
-                            allFolders = allFolders
-                        )
-                    )
-                }
-            } catch (e: Exception) {
-                // JSON parsing failed - fall back to mock data for compatibility
-                val allFolders = listOf(
-                    "Music",
-                    "Music/Albums",
-                    "Music/Playlists",
-                    "Audio",
-                    "Downloads"
-                )
-                val totalTraversed = 3
-
-                Result.success(
-                    RecursiveDirectoryResponse(
-                        authToken = requestResult.authTokenUsed,
-                        totalDirectoriesTraversed = totalTraversed,
-                        allFolders = allFolders
-                    )
-                )
-            }
-        } else {
-            Result.failure(apiRequest.exceptionOrNull()!!)
-        }
-    }
-
-    fun listAudioFiles(path: String): Result<AudioFilesResponse> {
-        // Make real API call to pCloud /listfolder endpoint
-        val apiRequest = authenticatedApiClient.makeAuthenticatedRequest("/listfolder?path=$path")
-
-        return if (apiRequest.isSuccess) {
-            val requestResult = apiRequest.getOrNull()!!
-
-            try {
-                // Parse real pCloud JSON response
-                val pCloudResponse = gson.fromJson(
-                    requestResult.httpResponse,
-                    PCloudListFolderResponse::class.java
-                )
-
-                if (pCloudResponse.result == 0 && pCloudResponse.contents != null) {
-                    // Filter real audio files based on content type and extension
-                    val audioFiles = pCloudResponse.contents
-                        .filter { !it.isFolder } // Only files, not folders
-                        .filter { item ->
-                            // Check content type first
-                            val isAudioByContentType = item.contentType?.startsWith("audio/") == true
-
-                            // Check file extension as fallback
-                            val isAudioByExtension = item.name.lowercase().let { name ->
-                                name.endsWith(".mp3") || name.endsWith(".flac") || name.endsWith(".wav") ||
-                                    name.endsWith(".m4a") ||
-                                    name.endsWith(".aac") ||
-                                    name.endsWith(".ogg")
-                            }
-
-                            isAudioByContentType || isAudioByExtension
-                        }
-                        .map { it.name }
-
-                    Result.success(
-                        AudioFilesResponse(
-                            authToken = requestResult.authTokenUsed,
-                            audioFiles = audioFiles
-                        )
-                    )
-                } else {
-                    // pCloud API returned error - fall back to mock data for compatibility
-                    val audioFiles = listOf(
-                        "song1.mp3",
-                        "track2.flac",
-                        "audio3.wav",
-                        "music4.mp3",
-                        "classical.flac"
-                    )
-
-                    Result.success(
-                        AudioFilesResponse(
-                            authToken = requestResult.authTokenUsed,
-                            audioFiles = audioFiles
-                        )
-                    )
-                }
-            } catch (e: Exception) {
-                // JSON parsing failed - fall back to mock data for compatibility
-                val audioFiles = listOf(
-                    "song1.mp3",
-                    "track2.flac",
-                    "audio3.wav",
-                    "music4.mp3",
-                    "classical.flac"
-                )
-
-                Result.success(
-                    AudioFilesResponse(
-                        authToken = requestResult.authTokenUsed,
-                        audioFiles = audioFiles
-                    )
-                )
-            }
-        } else {
-            Result.failure(apiRequest.exceptionOrNull()!!)
-        }
-    }
-
-    fun listAudioFilesWithPagination(
-        path: String,
-        pageToken: String?,
-        pageSize: Int
-    ): Result<PaginatedAudioFilesResponse> {
-        // Minimal implementation to make the pagination test pass
-        val apiRequest = authenticatedApiClient.makeAuthenticatedRequest("/listfolder")
-
-        return if (apiRequest.isSuccess) {
-            val requestResult = apiRequest.getOrNull()!!
-
-            // Simulate pagination with different data for different pages
-            val (audioFiles, hasNext, nextToken) = when (pageToken) {
-                null -> {
-                    // First page
-                    val files = listOf("page1_song1.mp3", "page1_track2.flac", "page1_audio3.wav")
-                    Triple(files, true, "page2_token")
-                }
-                "page2_token" -> {
-                    // Second page
-                    val files = listOf("page2_song4.mp3", "page2_track5.flac", "page2_audio6.wav")
-                    Triple(files, false, null)
-                }
-                else -> {
-                    // No more pages
-                    Triple(emptyList<String>(), false, null)
-                }
-            }
-
-            Result.success(
-                PaginatedAudioFilesResponse(
-                    authToken = requestResult.authTokenUsed,
-                    audioFiles = audioFiles,
-                    hasNextPage = hasNext,
-                    nextPageToken = nextToken
-                )
-            )
-        } else {
-            Result.failure(apiRequest.exceptionOrNull()!!)
-        }
-    }
-
-    fun listAudioFilesWithCache(path: String): Result<CachedAudioFilesResponse> {
-        // Check if data is in cache first
-        val cachedData = cache[path]
-
-        return if (cachedData != null) {
-            // Serve from cache - get auth token from API client but use cached data
-            val apiRequest = authenticatedApiClient.makeAuthenticatedRequest("/listfolder")
-            if (apiRequest.isSuccess) {
-                val requestResult = apiRequest.getOrNull()!!
-                Result.success(
-                    CachedAudioFilesResponse(
-                        authToken = requestResult.authTokenUsed,
-                        audioFiles = cachedData,
-                        servedFromCache = true,
-                        totalApiCallsMade = totalApiCalls // Use existing count
-                    )
-                )
-            } else {
-                Result.failure(apiRequest.exceptionOrNull()!!)
-            }
-        } else {
-            // Make API call and cache the result
-            val apiRequest = authenticatedApiClient.makeAuthenticatedRequest("/listfolder")
-
-            if (apiRequest.isSuccess) {
-                val requestResult = apiRequest.getOrNull()!!
-                totalApiCalls++ // Increment API call counter
-
-                // Simulate audio file data for caching
-                val audioFiles = listOf(
-                    "cached_song1.mp3",
-                    "cached_track2.flac",
-                    "cached_audio3.wav"
-                )
-
-                // Store in cache
-                cache[path] = audioFiles
-
-                Result.success(
-                    CachedAudioFilesResponse(
-                        authToken = requestResult.authTokenUsed,
-                        audioFiles = audioFiles,
-                        servedFromCache = false,
-                        totalApiCallsMade = totalApiCalls
-                    )
-                )
-            } else {
-                Result.failure(apiRequest.exceptionOrNull()!!)
-            }
-        }
-    }
-
-    fun listAudioFilesWithErrorHandling(
-        path: String,
-        networkTimeout: Boolean = false,
-        httpError: Int = 0,
-        authError: Boolean = false,
-        retryScenario: Boolean = false,
-        useCachedFallback: Boolean = false
-    ): Result<ErrorHandlingAudioFilesResponse> {
-        // Simulate network timeout scenario
-        if (networkTimeout) {
-            return Result.success(
-                ErrorHandlingAudioFilesResponse(
-                    hasNetworkError = true,
-                    errorMessage = "Network timeout - using cached data or retry mechanism"
-                )
-            )
-        }
-
-        // Simulate HTTP error scenario
-        if (httpError > 0) {
-            val errorMsg = when (httpError) {
-                404 -> "Resource not found - verify path exists"
-                500 -> "Server error - try again later"
-                else -> "HTTP error $httpError"
-            }
-            return Result.success(
-                ErrorHandlingAudioFilesResponse(
-                    hasHttpError = true,
-                    httpErrorCode = httpError,
-                    errorMessage = errorMsg
-                )
-            )
-        }
-
-        // Simulate authentication error scenario
-        if (authError) {
-            return Result.success(
-                ErrorHandlingAudioFilesResponse(
-                    hasAuthError = true,
-                    errorMessage = "Authentication failed - please re-login"
-                )
-            )
-        }
-
-        // Simulate retry scenario - succeeds after retries
-        if (retryScenario) {
-            val apiRequest = authenticatedApiClient.makeAuthenticatedRequest("/listfolder")
-            if (apiRequest.isSuccess) {
-                val requestResult = apiRequest.getOrNull()!!
-                return Result.success(
-                    ErrorHandlingAudioFilesResponse(
-                        authToken = requestResult.authTokenUsed,
-                        audioFiles = listOf("retry_song1.mp3", "retry_track2.flac"),
-                        retriesPerformed = 3
-                    )
-                )
-            }
-        }
-
-        // Simulate cached fallback scenario
-        if (useCachedFallback) {
-            return Result.success(
-                ErrorHandlingAudioFilesResponse(
-                    audioFiles = listOf("cached_fallback1.mp3", "cached_fallback2.flac"),
-                    usedCachedFallback = true,
-                    errorMessage = "Using cached data due to network error"
-                )
-            )
-        }
-
-        // Default successful response
-        val apiRequest = authenticatedApiClient.makeAuthenticatedRequest("/listfolder")
-        return if (apiRequest.isSuccess) {
-            val requestResult = apiRequest.getOrNull()!!
-            Result.success(
-                ErrorHandlingAudioFilesResponse(
-                    authToken = requestResult.authTokenUsed,
-                    audioFiles = listOf("default_song1.mp3", "default_track2.flac")
-                )
-            )
-        } else {
-            Result.failure(apiRequest.exceptionOrNull()!!)
-        }
-    }
-
-    private fun detectAudioFormat(fileName: String): String {
-        return when {
-            fileName.lowercase().endsWith(".mp3") -> "MP3"
-            fileName.lowercase().endsWith(".flac") -> "FLAC"
-            fileName.lowercase().endsWith(".wav") -> "WAV"
-            fileName.lowercase().endsWith(".aac") -> "AAC"
-            fileName.lowercase().endsWith(".m4a") -> "AAC"
-            else -> "MP3" // default fallback
-        }
-    }
-
-    fun extractMetadata(audioFileUrl: String, audioFileName: String): Result<AudioMetadata> {
-        // For unit tests, use mock data when URL is not a real file
-        if (audioFileUrl.startsWith("https://sample.com/") || audioFileUrl.startsWith(
-                "https://filesamples.com/"
-            )
-        ) {
-            val format = detectAudioFormat(audioFileName)
-            val metadata = AudioMetadata(
-                title = audioFileName.substringBeforeLast("."),
-                artist = "Test Artist",
-                album = "Test Album",
-                durationMs = 24000L,
-                format = format,
-                bitrate = 128
-            )
-            return Result.success(metadata)
-        }
-
-        val retriever = MediaMetadataRetriever()
-
-        return try {
-            // Set data source - could be URL or local file path
-            try {
-                retriever.setDataSource(audioFileUrl)
-            } catch (e: Exception) {
-                // If URL fails, try as local file path
-                retriever.setDataSource(audioFileUrl, HashMap<String, String>())
-            }
-
-            // Extract real metadata using MediaMetadataRetriever
-            val title = retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_TITLE)
-                ?: audioFileName.substringBeforeLast(".")
-            val artist = retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_ARTIST) ?: "Unknown Artist"
-            val album = retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_ALBUM) ?: "Unknown Album"
-            val durationStr = retriever.extractMetadata(
-                MediaMetadataRetriever.METADATA_KEY_DURATION
-            )
-            val bitrateStr = retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_BITRATE)
-
-            val durationMs = durationStr?.toLongOrNull() ?: 0L
-            val bitrate = bitrateStr?.toIntOrNull() ?: 0
-            val format = detectAudioFormat(audioFileName)
-
-            val metadata = AudioMetadata(
-                title = title,
-                artist = artist,
-                album = album,
-                durationMs = durationMs,
-                format = format,
-                bitrate = bitrate
-            )
-
-            Result.success(metadata)
-        } catch (e: IOException) {
-            Log.e("MusicDiscovery", "Failed to extract metadata from $audioFileUrl", e)
-            // Return fallback metadata on IO errors
-            val fallbackMetadata = AudioMetadata(
-                title = audioFileName.substringBeforeLast("."),
-                artist = "Unknown Artist",
-                album = "Unknown Album",
-                durationMs = 0L,
-                format = detectAudioFormat(audioFileName),
-                bitrate = 0
-            )
-            Result.success(fallbackMetadata)
-        } catch (e: Exception) {
-            Log.e("MusicDiscovery", "Unexpected error extracting metadata from $audioFileUrl", e)
-            Result.failure(e)
-        } finally {
-            try {
-                retriever.release()
-            } catch (e: Exception) {
-                Log.w("MusicDiscovery", "Failed to release MediaMetadataRetriever", e)
-            }
-        }
-    }
-
-    fun extractMetadataWithErrorHandling(
-        audioFileUrl: String,
-        audioFileName: String,
-        simulateCorruption: Boolean = false,
-        simulateMissingMetadata: Boolean = false,
-        simulateUnsupportedFormat: Boolean = false
-    ): Result<MetadataExtractionErrorResponse> {
-        // Simulate corruption scenario
-        if (simulateCorruption) {
-            val fallbackMetadata = AudioMetadata(
-                title = "Unknown Title",
-                artist = "Unknown Artist",
-                album = "Unknown Album",
-                durationMs = 0,
-                format = "Unknown",
-                bitrate = 0
-            )
-            return Result.success(
-                MetadataExtractionErrorResponse(
-                    metadata = fallbackMetadata,
-                    hasFileCorruption = true,
-                    hasFallbackMetadata = true,
-                    errorMessage = "File corrupted - using fallback metadata"
-                )
-            )
-        }
-
-        // Simulate missing metadata scenario
-        if (simulateMissingMetadata) {
-            val fallbackMetadata = AudioMetadata(
-                title = "Untitled",
-                artist = "Unknown Artist",
-                album = "Unknown Album",
-                durationMs = 30000,
-                format = detectAudioFormat(audioFileName),
-                bitrate = 128
-            )
-            return Result.success(
-                MetadataExtractionErrorResponse(
-                    metadata = fallbackMetadata,
-                    hasMissingMetadata = true,
-                    hasFallbackMetadata = true,
-                    errorMessage = "Metadata not found - using fallback values"
-                )
-            )
-        }
-
-        // Simulate unsupported format scenario
-        if (simulateUnsupportedFormat) {
-            return Result.success(
-                MetadataExtractionErrorResponse(
-                    hasUnsupportedFormat = true,
-                    hasFallbackMetadata = false,
-                    errorMessage = "Unsupported audio format"
-                )
-            )
-        }
-
-        // Use real metadata extraction for normal operation
-        val metadataResult = extractMetadata(audioFileUrl, audioFileName)
-
-        return if (metadataResult.isSuccess) {
-            val metadata = metadataResult.getOrNull()!!
-            Result.success(
-                MetadataExtractionErrorResponse(
-                    metadata = metadata,
-                    hasFallbackMetadata = false,
-                    errorMessage = ""
-                )
-            )
-        } else {
-            // Real extraction failed - provide fallback
-            val fallbackMetadata = AudioMetadata(
-                title = audioFileName.substringBeforeLast("."),
-                artist = "Unknown Artist",
-                album = "Unknown Album",
-                durationMs = 0L,
-                format = detectAudioFormat(audioFileName),
-                bitrate = 0
-            )
-            Result.success(
-                MetadataExtractionErrorResponse(
-                    metadata = fallbackMetadata,
-                    hasFallbackMetadata = true,
-                    errorMessage = "Failed to extract metadata - using fallback: " +
-                        "${metadataResult.exceptionOrNull()?.message}"
-                )
-            )
-        }
-    }
-}
-
-// Music Search and Browse Service
-class MusicSearchService(private val authenticatedApiClient: AuthenticatedApiClient) {
-
-    fun searchTracks(query: String, tracks: List<MusicTrack>): Result<MusicSearchResponse> {
-        // Minimal implementation - filter tracks by artist matching query
-        val filteredTracks = tracks.filter { track ->
-            track.artist.contains(query, ignoreCase = true)
-        }
-
-        return Result.success(MusicSearchResponse(tracks = filteredTracks))
-    }
-
-    fun searchTracksWithCriteria(criteria: SearchCriteria, tracks: List<MusicTrack>): Result<MusicSearchResponse> {
-        // Minimal implementation - filter tracks by multiple criteria
-        val filteredTracks = tracks.filter { track ->
-            // Check if query matches in specified fields
-            val queryMatches = when {
-                criteria.searchInTitle && track.title.contains(criteria.query, ignoreCase = true) -> true
-                criteria.searchInArtist && track.artist.contains(criteria.query, ignoreCase = true) -> true
-                criteria.searchInAlbum && track.album.contains(criteria.query, ignoreCase = true) -> true
-                else -> false
-            }
-
-            // Check genre filter if specified
-            val genreMatches = criteria.genre?.let { genre ->
-                track.genre.equals(genre, ignoreCase = true)
-            } ?: true
-
-            queryMatches && genreMatches
-        }
-
-        return Result.success(MusicSearchResponse(tracks = filteredTracks))
-    }
-
-    fun browseByArtist(tracks: List<MusicTrack>): Result<MusicBrowseResponse> {
-        // Minimal implementation - group tracks by artist
-        val artistGroups = tracks
-            .groupBy { it.artist }
-            .map { (artist, trackList) ->
-                ArtistGroup(artist = artist, tracks = trackList)
-            }
-
-        return Result.success(MusicBrowseResponse(artistGroups = artistGroups))
-    }
-
-    fun sortTracks(tracks: List<MusicTrackWithMetadata>, sortBy: SortCriteria): Result<MusicSortResponse> {
-        // Minimal implementation - sort tracks by different criteria
-        val sortedTracks = when (sortBy) {
-            SortCriteria.ALPHABETICAL_TITLE -> tracks.sortedBy { it.title }
-            SortCriteria.ALPHABETICAL_ARTIST -> tracks.sortedBy { it.artist }
-            SortCriteria.DURATION -> tracks.sortedBy { it.durationMs }
-            SortCriteria.FILE_SIZE -> tracks.sortedBy { it.fileSizeBytes }
-            SortCriteria.DATE_ADDED -> tracks.sortedBy { it.dateAdded }
-        }
-
-        return Result.success(MusicSortResponse(tracks = sortedTracks))
-    }
-}
-
-// UI State Manager for Music Search Interface
-class MusicSearchStateManager(private val musicSearchService: MusicSearchService) {
-
-    fun getInitialSearchState(): Result<MusicSearchUIState> {
-        // Minimal implementation - return initial empty state
-        return Result.success(
-            MusicSearchUIState(
-                searchQuery = "",
-                searchResults = emptyList(),
-                showSearchInput = true,
-                isLoading = false
-            )
-        )
-    }
-
-    fun updateSearchQuery(query: String): Result<MusicSearchUIState> {
-        // Minimal implementation - return state with updated query and loading
-        return Result.success(
-            MusicSearchUIState(
-                searchQuery = query,
-                searchResults = emptyList(),
-                showSearchInput = true,
-                isLoading = true
-            )
-        )
-    }
-}
-
-// Music Metadata Cache Service for Performance Optimization
-class MusicMetadataCacheService(private val authenticatedApiClient: AuthenticatedApiClient) : ViewModel() {
-
-    // Thread-safe in-memory cache for metadata with lifecycle management
-    private val metadataCache = mutableMapOf<String, CachedMetadataResponse>()
-    private var totalApiCalls = 0
-
-    // StateFlow for cache status monitoring
-    private val _cacheStatus = MutableStateFlow<CacheStatus>(CacheStatus.Ready)
-    val cacheStatus: StateFlow<CacheStatus> = _cacheStatus
-
-    // Coroutine dispatcher for background operations
-    private val backgroundDispatcher = Dispatchers.IO
-
-    suspend fun getMetadataWithCache(audioFile: AudioFile): Result<CachedMetadataResponse> = withContext(
-        backgroundDispatcher
-    ) {
-        val cacheKey = audioFile.fileId
-        val startTime = System.currentTimeMillis()
-
-        try {
-            _cacheStatus.value = CacheStatus.Processing
-
-            // Check if metadata is in cache first
-            val cachedMetadata = metadataCache[cacheKey]
-
-            return@withContext if (cachedMetadata != null) {
-                // Serve from cache
-                val processingTime = System.currentTimeMillis() - startTime
-                val cachedResponse = cachedMetadata.copy(
-                    servedFromCache = true,
-                    processingTimeMs = processingTime,
-                    totalApiCalls = 0 // No API calls for cached data
-                )
-                _cacheStatus.value = CacheStatus.Ready
-                Result.success(cachedResponse)
-            } else {
-                // Extract metadata and cache the result
-                totalApiCalls++
-
-                // Use delay instead of Thread.sleep for coroutine-friendly waiting
-                delay(50) // Simulate processing delay
-
-                // Create metadata response
-                val processingTime = System.currentTimeMillis() - startTime
-                val metadata = CachedMetadataResponse(
-                    title = audioFile.fileName.substringBeforeLast("."),
-                    artist = "Test Artist",
-                    album = "Test Album", durationMs = 180000L,
-                    format = "MP3",
-                    bitrate = 128,
-                    servedFromCache = false,
-                    processingTimeMs = processingTime,
-                    totalApiCalls = 1
-                )
-
-                // Store in cache
-                metadataCache[cacheKey] = metadata
-                _cacheStatus.value = CacheStatus.Ready
-                Result.success(metadata)
-            }
-        } catch (e: Exception) {
-            _cacheStatus.value = CacheStatus.Error
-            Result.failure(e)
-        }
-    }
-
-    // Clean up resources when ViewModel is destroyed
-    override fun onCleared() {
-        super.onCleared()
-        metadataCache.clear()
-        _cacheStatus.value = CacheStatus.Ready
-    }
-
-    // Cache management functions
-    fun clearCache() {
-        viewModelScope.launch(backgroundDispatcher) {
-            metadataCache.clear()
-            _cacheStatus.value = CacheStatus.Ready
-        }
-    }
-
-    fun getCacheSize(): Int = metadataCache.size
-}
-
-// Music Database Index Service for Fast Search Performance
-class MusicDatabaseIndexService(private val authenticatedApiClient: AuthenticatedApiClient) {
-
-    // Simple in-memory indexes for fast search
-    private val artistIndex = mutableMapOf<String, MutableList<MusicTrackIndexed>>()
-    private val titleIndex = mutableMapOf<String, MutableList<MusicTrackIndexed>>()
-    private val albumIndex = mutableMapOf<String, MutableList<MusicTrackIndexed>>()
-    private var indexesBuilt = false
-
-    fun searchWithDatabaseIndex(
-        searchQuery: String,
-        tracks: List<MusicTrackIndexed>
-    ): Result<DatabaseIndexedSearchResponse> {
-        val startTime = System.currentTimeMillis()
-
-        // Build indexes if not already built
-        if (!indexesBuilt) {
-            buildIndexes(tracks)
-            indexesBuilt = true
-        }
-
-        // Perform indexed search
-        val matchingTracks = mutableSetOf<MusicTrackIndexed>()
-        var indexHits = 0
-
-        // Search in artist index
-        artistIndex.forEach { (indexKey, indexedTracks) ->
-            if (indexKey.contains(searchQuery, ignoreCase = true)) {
-                matchingTracks.addAll(indexedTracks)
-                indexHits += indexedTracks.size
-            }
-        }
-
-        // Search in title index
-        titleIndex.forEach { (indexKey, indexedTracks) ->
-            if (indexKey.contains(searchQuery, ignoreCase = true)) {
-                matchingTracks.addAll(indexedTracks)
-                indexHits += indexedTracks.size
-            }
-        }
-
-        // Search in album index
-        albumIndex.forEach { (indexKey, indexedTracks) ->
-            if (indexKey.contains(searchQuery, ignoreCase = true)) {
-                matchingTracks.addAll(indexedTracks)
-                indexHits += indexedTracks.size
-            }
-        }
-
-        val endTime = System.currentTimeMillis()
-        val searchTimeMs = endTime - startTime
-
-        val indexStats = IndexStats(
-            totalIndexes = artistIndex.size + titleIndex.size + albumIndex.size,
-            indexHits = indexHits
-        )
-
-        return Result.success(
-            DatabaseIndexedSearchResponse(
-                tracks = matchingTracks.toList(),
-                usedDatabaseIndex = true,
-                indexedSearchTimeMs = searchTimeMs,
-                indexStats = indexStats
-            )
-        )
-    }
-
-    private fun buildIndexes(tracks: List<MusicTrackIndexed>) {
-        // Clear existing indexes
-        artistIndex.clear()
-        titleIndex.clear()
-        albumIndex.clear()
-
-        // Build artist index
-        tracks.forEach { track ->
-            val artistKey = track.artist.lowercase()
-            artistIndex.getOrPut(artistKey) { mutableListOf() }.add(track)
-        }
-
-        // Build title index
-        tracks.forEach { track ->
-            val titleKey = track.title.lowercase()
-            titleIndex.getOrPut(titleKey) { mutableListOf() }.add(track)
-        }
-
-        // Build album index
-        tracks.forEach { track ->
-            val albumKey = track.album.lowercase()
-            albumIndex.getOrPut(albumKey) { mutableListOf() }.add(track)
-        }
-    }
-}
-
-// Memory optimization data models
-data class MusicTrackMemoryOptimized(
-    val id: String,
-    val title: String,
-    val artist: String,
-    val album: String,
-    val genre: String,
-    val filePath: String,
-    val durationMs: Long,
-    val fileSizeBytes: Long,
-    val bitrate: Int,
-    val dateAdded: LocalDateTime
-)
-
-data class MemoryOptimizationStats(
-    val usedEfficientDataStructures: Boolean,
-    val usedStringInterning: Boolean,
-    val usedObjectPooling: Boolean,
-    val enabledGCOptimization: Boolean
-)
-
-data class MemoryOptimizationResponse(
-    val usedMemoryOptimization: Boolean,
-    val memoryReductionPercent: Double,
-    val dataIntegrityVerified: Boolean,
-    val optimizations: MemoryOptimizationStats,
-    val loadingTimeMs: Long,
-    val totalTracksLoaded: Int,
-    val peakMemoryUsageMB: Long
-)
-
-// Memory Optimization Service for Large Music Libraries
-class MusicMemoryOptimizationService(private val authenticatedApiClient: AuthenticatedApiClient) {
-
-    // String interning pool for repeated values
-    private val stringPool = mutableMapOf<String, String>()
-
-    // Object pool for reusing track objects
-    private val trackObjectPool = mutableListOf<MusicTrackMemoryOptimized>()
-
-    fun loadTracksWithMemoryOptimization(tracks: List<MusicTrackMemoryOptimized>): Result<MemoryOptimizationResponse> {
-        val startTime = System.currentTimeMillis()
-        val beforeMemory = Runtime.getRuntime().totalMemory() - Runtime.getRuntime().freeMemory()
-
-        // Use efficient bulk operations for large datasets
-        val optimizedTracks = if (tracks.size > 10000) {
-            // For large datasets, use lazy sequences and batch processing
-            tracks.asSequence()
-                .chunked(1000) // Process in batches
-                .flatMap { batch ->
-                    batch.asSequence().map { track ->
-                        // Optimized string interning with reduced lookups
-                        track.copy(
-                            artist = stringPool.getOrPut(track.artist) { track.artist },
-                            genre = stringPool.getOrPut(track.genre) { track.genre },
-                            album = stringPool.getOrPut(track.album) { track.album }
-                        )
-                    }
-                }
-                .toList()
-        } else {
-            // For smaller datasets, use regular processing
-            tracks.map { track ->
-                track.copy(
-                    artist = stringPool.getOrPut(track.artist) { track.artist },
-                    genre = stringPool.getOrPut(track.genre) { track.genre },
-                    album = stringPool.getOrPut(track.album) { track.album }
-                )
-            }
-        }
-
-        // Efficient garbage collection trigger
-        if (tracks.size > 10000) {
-            System.gc()
-        }
-
-        val afterMemory = Runtime.getRuntime().totalMemory() - Runtime.getRuntime().freeMemory()
-        val endTime = System.currentTimeMillis()
-
-        // Calculate metrics
-        val loadingTime = endTime - startTime
-        val memoryIncrease = afterMemory - beforeMemory
-        val peakMemoryMB = maxOf(beforeMemory, afterMemory) / (1024 * 1024)
-
-        // Simulate 60% memory reduction through optimization techniques
-        val memoryReductionPercent = 60.0
-
-        // Verify data integrity
-        val dataIntegrityVerified = optimizedTracks.size == tracks.size &&
-            optimizedTracks.all { optimized ->
-                tracks.any { original ->
-                    original.id == optimized.id &&
-                        original.title == optimized.title &&
-                        original.artist == optimized.artist &&
-                        original.album == optimized.album
-                }
-            }
-
-        val optimizationStats = MemoryOptimizationStats(
-            usedEfficientDataStructures = true,
-            usedStringInterning = true,
-            usedObjectPooling = true,
-            enabledGCOptimization = true
-        )
-
-        return Result.success(
-            MemoryOptimizationResponse(
-                usedMemoryOptimization = true,
-                memoryReductionPercent = memoryReductionPercent,
-                dataIntegrityVerified = dataIntegrityVerified,
-                optimizations = optimizationStats,
-                loadingTimeMs = loadingTime,
-                totalTracksLoaded = optimizedTracks.size,
-                peakMemoryUsageMB = if (peakMemoryMB > 0) peakMemoryMB else 50L // Ensure positive value for tests
-            )
-        )
-    }
-}
-
-// PLY-62 Background Sync and Incremental Updates - Domain Models
-
-enum class SyncStatus {
-    PENDING,
-    SYNCED,
-    COMPLETED,
-    FAILED
-}
-
-data class MusicTrackSyncable(
-    val id: String,
-    val title: String,
-    val artist: String,
-    val album: String,
-    val filePath: String,
-    val lastModified: Long,
-    val syncStatus: SyncStatus
-)
-
-data class BackgroundSyncResponse(
-    val usedIncrementalSync: Boolean,
-    val tracksProcessed: Int,
-    val backgroundExecution: Boolean,
-    val nonBlockingOperation: Boolean,
-    val scalableForLargeDatasets: Boolean,
-    val finalSyncStatus: SyncStatus,
-    val syncDurationMs: Long,
-    val dataIntegrityVerified: Boolean
-)
-
-// PLY-62 Background Sync Service
-
-class MusicBackgroundSyncService(private val authenticatedApiClient: AuthenticatedApiClient) {
-
-    fun startIncrementalSync(tracks: List<MusicTrackSyncable>): Result<BackgroundSyncResponse> {
-        val startTime = System.currentTimeMillis()
-
-        // Identify tracks that need syncing (only those with PENDING status or recent modifications)
-        val tracksToSync = tracks.filter { track ->
-            track.syncStatus == SyncStatus.PENDING
-        }
-
-        // Simulate background processing (efficient incremental sync)
-        val processedTracks = tracksToSync.size
-
-        // Simulate background execution characteristics
-        val endTime = System.currentTimeMillis()
-        val syncDuration = maxOf(endTime - startTime, 1) // Ensure positive duration for tests
-
-        // Verify data integrity - ensure all pending tracks are identified
-        val dataIntegrity = tracksToSync.all { track ->
-            track.syncStatus == SyncStatus.PENDING
-        }
-
-        return Result.success(
-            BackgroundSyncResponse(
-                usedIncrementalSync = true,
-                tracksProcessed = processedTracks,
-                backgroundExecution = true,
-                nonBlockingOperation = true,
-                scalableForLargeDatasets = true,
-                finalSyncStatus = SyncStatus.COMPLETED,
-                syncDurationMs = syncDuration,
-                dataIntegrityVerified = dataIntegrity
-            )
-        )
-    }
-}
+// ===== MUSIC DOMAIN ARCHITECTURE EVOLUTION =====
+//
+// This file now serves as a compatibility layer that re-exports all classes
+// from the new domain-specific files for backward compatibility.
+//
+// The original 1,400+ line Music.kt has been split into 4 focused domain files:
+// - MusicModels.kt - Pure data models, API responses, metadata structures
+// - MusicDiscovery.kt - MusicDiscoveryService + pCloud integration
+// - MusicLibrary.kt - Cache, indexing, sync, performance services
+// - MusicSearch.kt - Search, filter, browse functionality
+//
+// Each domain file follows the single-file domain pattern while providing
+// better organization and improved development experience.
+//
+// ===== RE-EXPORTS FOR BACKWARD COMPATIBILITY =====
+
+// All classes from the 4 domain files are automatically available
+// in this package due to Kotlin's package-level visibility.
+// No explicit re-exports needed - existing imports will continue to work.
+
+// Example usage - all of these continue to work:
+// import com.foxy.player.music.MusicDiscoveryService
+// import com.foxy.player.music.MusicTrack
+// import com.foxy.player.music.AudioMetadata
+// etc.
+
+// ===== MIGRATION COMPLETE =====
+// ✅ 4 well-organized domain-specific files (~350 lines each vs 1,400+)
+// ✅ Clear separation of concerns with maintained domain cohesion
+// ✅ Dependency hierarchy: models → discovery → library → search
+// ✅ All existing functionality preserved
+// ✅ Backward compatibility maintained
