+++ app/src/main/java/com/foxy/player/music/MusicSearch.kt
@@ -1,0 +1,94 @@
+package com.foxy.player.music
+
+import com.foxy.player.authentication.AuthenticatedApiClient
+
+// ===== MUSIC SEARCH AND BROWSE SERVICE =====
+
+// Music Search and Browse Service
+class MusicSearchService(private val authenticatedApiClient: AuthenticatedApiClient) {
+
+    fun searchTracks(query: String, tracks: List<MusicTrack>): Result<MusicSearchResponse> {
+        // Minimal implementation - filter tracks by artist matching query
+        val filteredTracks = tracks.filter { track ->
+            track.artist.contains(query, ignoreCase = true)
+        }
+
+        return Result.success(MusicSearchResponse(tracks = filteredTracks))
+    }
+
+    fun searchTracksWithCriteria(criteria: SearchCriteria, tracks: List<MusicTrack>): Result<MusicSearchResponse> {
+        // Minimal implementation - filter tracks by multiple criteria
+        val filteredTracks = tracks.filter { track ->
+            // Check if query matches in specified fields
+            val queryMatches = when {
+                criteria.searchInTitle && track.title.contains(criteria.query, ignoreCase = true) -> true
+                criteria.searchInArtist && track.artist.contains(criteria.query, ignoreCase = true) -> true
+                criteria.searchInAlbum && track.album.contains(criteria.query, ignoreCase = true) -> true
+                else -> false
+            }
+
+            // Check genre filter if specified
+            val genreMatches = criteria.genre?.let { genre ->
+                track.genre.equals(genre, ignoreCase = true)
+            } ?: true
+
+            queryMatches && genreMatches
+        }
+
+        return Result.success(MusicSearchResponse(tracks = filteredTracks))
+    }
+
+    fun browseByArtist(tracks: List<MusicTrack>): Result<MusicBrowseResponse> {
+        // Minimal implementation - group tracks by artist
+        val artistGroups = tracks
+            .groupBy { it.artist }
+            .map { (artist, trackList) ->
+                ArtistGroup(artist = artist, tracks = trackList)
+            }
+
+        return Result.success(MusicBrowseResponse(artistGroups = artistGroups))
+    }
+
+    fun sortTracks(tracks: List<MusicTrackWithMetadata>, sortBy: SortCriteria): Result<MusicSortResponse> {
+        // Minimal implementation - sort tracks by different criteria
+        val sortedTracks = when (sortBy) {
+            SortCriteria.ALPHABETICAL_TITLE -> tracks.sortedBy { it.title }
+            SortCriteria.ALPHABETICAL_ARTIST -> tracks.sortedBy { it.artist }
+            SortCriteria.DURATION -> tracks.sortedBy { it.durationMs }
+            SortCriteria.FILE_SIZE -> tracks.sortedBy { it.fileSizeBytes }
+            SortCriteria.DATE_ADDED -> tracks.sortedBy { it.dateAdded }
+        }
+
+        return Result.success(MusicSortResponse(tracks = sortedTracks))
+    }
+}
+
+// ===== UI STATE MANAGER FOR MUSIC SEARCH =====
+
+// UI State Manager for Music Search Interface
+class MusicSearchStateManager(private val musicSearchService: MusicSearchService) {
+
+    fun getInitialSearchState(): Result<MusicSearchUIState> {
+        // Minimal implementation - return initial empty state
+        return Result.success(
+            MusicSearchUIState(
+                searchQuery = "",
+                searchResults = emptyList(),
+                showSearchInput = true,
+                isLoading = false
+            )
+        )
+    }
+
+    fun updateSearchQuery(query: String): Result<MusicSearchUIState> {
+        // Minimal implementation - return state with updated query and loading
+        return Result.success(
+            MusicSearchUIState(
+                searchQuery = query,
+                searchResults = emptyList(),
+                showSearchInput = true,
+                isLoading = true
+            )
+        )
+    }
+}
